+++ src/react-components/CHANGELOG.md
@@ -7,8 +7,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ========
 
 ## [0.3.6]
-
--   added `Template` stories
+- added `Template` stories
 
 ## [0.3.5]